AiAdventChallenge/Day1/tgAiBot/requirements.txt
python-telegram-bot==22.5
openai==2.15.0
logging==0.4.9.6